refactor - todo list memo
